[MIN-50] Separate runtime tests into modules
